+++ data/priv/repo/migrations/20250328060063_create_updates.exs
@@ -5,13 +5,13 @@ defmodule Moc.Data.Repo.Migrations.CreateUpdates do
     create table("updates") do
       add :type, :string
       add :xp, :float, null: true
-      add :level, :integer, null: true
+      add :level, :float, null: true
       add :title, :string, null: true
       add :prefix, :string, null: true
-      add :charisma, :integer, null: true
-      add :constitution, :integer, null: true
-      add :dexterity, :integer, null: true
-      add :wisdom, :integer, null: true
+      add :charisma, :float, null: true
+      add :constitution, :float, null: true
+      add :dexterity, :float, null: true
+      add :wisdom, :float, null: true
       add :medal_id, references("medals", on_delete: :nothing), null: true
       add :contributor_id, references("contributors", on_delete: :delete_all), null: false
